+++ bricks/flutter_starter_brick/__brick__/lib/features/random_jokes/src/models/ui_models/ui_joke.dart
@@ -1,9 +1,6 @@
 /// UI model for displaying a joke
 final class UiJoke {
-  const UiJoke({
-    required this.id,
-    required this.content,
-  });
+  const UiJoke({required this.id, required this.content});
 
   /// Unique identifier for the joke
   final String id;