+++ bricks/flutter_starter_brick/__brick__/lib/features/random_jokes/src/ui/joke_card.dart
@@ -5,22 +5,19 @@ import 'package:{{proj_name}}/foundation/ui/widgets/widgets.dart';
 
 /// Card widget for displaying a joke
 class JokeCard extends StatelessWidget {
-  const JokeCard({
-    required this.joke,
-    super.key,
-  });
+  const JokeCard({required this.joke, super.key});
 
   final UiJoke joke;
 
   @override
   Widget build(BuildContext context) => Card(
-    margin: EdgeInsets.all(SpacingSize.medium.value),
+    margin: EdgeInsets.all(SpacingSize.spacing24.value),
     elevation: 2,
     child: Padding(
-      padding: EdgeInsets.all(SpacingSize.large.value),
+      padding: EdgeInsets.all(SpacingSize.spacing32.value),
       child: Text(
         joke.content,
-        style: context.typography?.body2,
+        style: context.typography?.bodyText,
         textAlign: TextAlign.center,
       ),
     ),