+++ bricks/flutter_starter_brick/__brick__/lib/foundation/ui/overlays/src/dialog_with_two_actions.dart
@@ -31,7 +31,12 @@ Future<T?> showDialogWithTwoActions<T>({
     title: Padding(
       padding: const EdgeInsets.only(top: DialogConstants.verticalPadding),
       child: Center(
-        child: Text(title, style: context.typography?.title6),
+        child: Text(
+          title,
+          style: context.typography?.primaryTitle.copyWith(
+            color: context.themeData.colorScheme.onSurface,
+          ),
+        ),
       ),
     ),
     content: Padding(
@@ -59,15 +64,16 @@ Future<T?> showDialogWithTwoActions<T>({
           ),
           const SizedBox(width: DialogConstants.buttonSpacing),
           Expanded(
-            child: rightButtonIsDestructive
-                ? DestructiveButton(
-                    text: rightButtonText.toUpperCase(),
-                    onPressed: rightButtonOnPressed,
-                  )
-                : MainButton(
-                    text: rightButtonText.toUpperCase(),
-                    onPressed: rightButtonOnPressed,
-                  ),
+            child:
+                rightButtonIsDestructive
+                    ? DestructiveButton(
+                      text: rightButtonText.toUpperCase(),
+                      onPressed: rightButtonOnPressed,
+                    )
+                    : MainButton(
+                      text: rightButtonText.toUpperCase(),
+                      onPressed: rightButtonOnPressed,
+                    ),
           ),
         ],
       ),