+++ bricks/flutter_starter_brick/__brick__/lib/foundation/ui/theme/src/theme_data/text_button_theme_data.dart
@@ -1,4 +1,4 @@
-import 'package:flutter/material.dart';
+import 'package:flutter/material.dart' hide Typography;
 import 'package:{{proj_name}}/foundation/ui/theme/src/theme_data/defaults.dart';
 import 'package:{{proj_name}}/foundation/ui/theme/src/theme_data/typography.dart';
 import 'package:{{proj_name}}/resources/resources.dart';
@@ -24,15 +24,21 @@ TextButtonThemeData textButtonThemeData(ColorScheme colorScheme) =>
         ),
         overlayColor: WidgetStateProperty.resolveWith((states) {
           if (states.contains(WidgetState.disabled)) return null;
-          return colorScheme.primary.withValues(alpha: ThemeDefaults.buttonDisabledOverlayAlpha);
+          return colorScheme.primary.withValues(
+            alpha: ThemeDefaults.buttonDisabledOverlayAlpha,
+          );
         }),
         elevation: const WidgetStatePropertyAll(0),
         textStyle: WidgetStateProperty.all(
-          Fonts.defaultTitleFont.textStyle(fontSize: ThemeDefaults.buttonTextSize),
+          Typography.defaultButtonFont.textStyle.copyWith(
+            fontSize: ThemeDefaults.buttonTextSize,
+          ),
         ),
         foregroundColor: WidgetStateProperty.resolveWith((states) {
           if (states.contains(WidgetState.disabled)) {
-            return colorScheme.onSurface.withValues(alpha: ThemeDefaults.materialDisabledOpacity);
+            return colorScheme.onSurface.withValues(
+              alpha: ThemeDefaults.materialDisabledOpacity,
+            );
           }
           return colorScheme.primary;
         }),