+++ bricks/flutter_starter_brick/__brick__/lib/foundation/ui/widgets/src/blur_widget.dart
@@ -21,15 +21,16 @@ class BlurWidget extends StatelessWidget {
   final double blurIntensity;
 
   @override
-  Widget build(BuildContext context) => applyBlur
-      ? BackdropFilter(
-          filter: ImageFilter.blur(
-            sigmaX: blurIntensity,
-            sigmaY: blurIntensity,
-          ),
-          child: child,
-        )
-      : child;
+  Widget build(BuildContext context) =>
+      applyBlur
+          ? BackdropFilter(
+            filter: ImageFilter.blur(
+              sigmaX: blurIntensity,
+              sigmaY: blurIntensity,
+            ),
+            child: child,
+          )
+          : child;
 }
 
 class BlurWidgetDefaults {